Fix up the colours
On bash (4.4 at least) echo needs -e to interpret the escape equences in
the string, otherwise it just displays it as is.

On the other hand, the sh compatible shell 'dash' does not understand
echo -e nor echo -ne so for simplicity's sake the shebang was switched
to depend on bash directly, which supports both.